chore(api): move solution-service-qrocm-dep.js to evaluation scope
