+++ api/lib/infrastructure/repositories/index.js
@@ -3,9 +3,9 @@ import * as trainingTriggerRepository from '../../../src/devcomp/infrastructure/
 import * as tutorialEvaluationRepository from '../../../src/devcomp/infrastructure/repositories/tutorial-evaluation-repository.js';
 import * as tutorialRepository from '../../../src/devcomp/infrastructure/repositories/tutorial-repository.js';
 import * as userRecommendedTrainingRepository from '../../../src/devcomp/infrastructure/repositories/user-recommended-training-repository.js';
+import { getCorrection } from '../../../src/evaluation/domain/services/solution/solution-service-qrocm-dep.js';
 import { fromDatasourceObject } from '../../../src/shared/infrastructure/adapters/solution-adapter.js';
 import { injectDependencies } from '../../../src/shared/infrastructure/utils/dependency-injection.js';
-import { getCorrection } from '../../domain/services/solution-service-qrocm-dep.js';
 import * as correctionRepository from './correction-repository.js';
 
 /**
